MAT-6884 error checking when saving test case with special characters
