+++ src/util/checkSpecialCharacters.test.tsx
